[wip] refs add-apollo-local-state
